rename: [조안나] 사용자가 판매자로 로그인 시 모든 페이지에서 사용되는 Common App Bar UI [DIY-SALES-APP-UI 19]
